Add initial scripts for data reduction and visualization
This commit introduces a script (`reduce-transaction.py`) for reducing dataset size based on variance and a script (`visualisation_V1-V2.py`) for visualizing principal components V1 and V2 with seaborn. Additionally, a `requirements.txt` file has been added to list dependencies, and a basic `main.py` script for loading the dataset is included.